Rename to `rubyevents`
