+++ .rubocop-custom.yml
@@ -1,5 +1,5 @@
 require:
-  - ./lib/rubocop/cop/ruby_events/prefer_yerba.rb
+  - ./lib/rubocop/cop/rubyevents/prefer_yerba.rb
 
 RubyEvents/PreferYerba:
   Enabled: true
